requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
idna==2.10
lxml==4.6.2
python-docx==0.8.10
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
xlwt==1.3.0
